refactor: Change dags structure
