+++ airflow/dags/.airflowignore
@@ -1,0 +1,2 @@
+//**
+!/pipe/
