--- configs/config.json
@@ -1,26 +1,0 @@
-{
-  "projeto": {
-    "nome": "Ingestor Star Wars API",
-    "versao": "1.0.0"
-  },
-  "tabelas": [
-    {
-      "name": "people",
-      "label": "Personagens",
-      "url": "https://swapi.dev/api/people?page={page}",
-      "path": "./data/people/{datetime}_people.json"
-    },
-    {
-      "name": "planets",
-      "label": "Planetas",
-      "url": "https://swapi.dev/api/planets?page={page}",
-      "path": "./data/planets/{datetime}_planets.json"
-    },
-    {
-      "name": "films",
-      "label": "Filmes",
-      "url": "https://swapi.dev/api/films?page={page}",
-      "path": "./data/films/{datetime}_films.json"
-    }
-  ]
-}
